fix: address download flow review comments
