+++ src/components/ReleasesTabs.tsx
@@ -9,9 +9,8 @@ class ReleasesTabs extends React.Component {
     platform: string;
     releases: {
       name: string;
-      id: string;
       title: string;
-      downloads: DownloadItem[];
+      downloads?: DownloadItem[] | null;
       description: string;
     }[];
   }) {
